Closure Coupon Refund
One that I specifically asked for progress on was the issue with coupons not being refunded, even when processing an admin refund all request. I tried processing a refund for the entire amount to a coupon and only the portion paid above and beyond the coupon value was refunded (see below).