1859 Add CAPTCHA on both GPXVacations.com and My-GPX.com
Reset Resources.
This work was performed by Chris Goering. It's probably obvious since you can see who submitted this. Traci was led to believe by Mark that someone else would be doing the work. This delayed the other work that Traci anticipated being done making me look like I'm not doing a good enough job once again. In fact, Mark told another developer that he was going to fire me because I couldn't get the work done on time. The real problem is that he wouldn't be honest with the client and tell them what was really assigned to me. For example, I quoted VEST as if we were going to pay a subcontractor to do the work and I would just manage them. That subcontractor didn't do their job and Mark fired them. Then he told me that I was responsible all of the development work which delayed other work that I was responsible for. In addition, he only offered a few thousand dollars to compensate me for my time even though I ended up averaging forty hours per week on VEST and I ended up making about $12 per hour which isn't even a living wage when I'm responsible for the majority of the income. Back to the time issue. I was unable to complete other work that I needed to do because I was forced to work on VEST which delayed every other project. I was trying to catch up while also getting GPX work done making it impossible to meet previously set deadlines. My work began to suffer because I was both rushing as well as working too many hours. The longer the day the more likely I was to make a mistake. So I try to work faster and also make a mistake. But he wants to fire me because I'm not getting the work done. As you can guess, I cannot continue to work in this hostile environment and I will ending the relationship as soon as I can break away from the commitments that I have already made to him. Contact chris@brandmediasolutions.com with questions/concerns.